requirements.txt
fastapi == 0.89.1
langchain==0.0.351
sentence-transformers==2.7.0
torch==2.0.1
uvicorn==0.20.0
python-dotenv==1.0.0
cryptography==42.0.2
